backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-multipart
pytesseract
pillow
httpx